refactor: 모달 밖 영역 클릭 function utils 폴더에 분리
